+++ web/src/nav/GlobalNavbar.tsx
@@ -5,9 +5,7 @@ import { Subscription } from 'rxjs'
 import { authRequired } from '../auth'
 import * as GQL from '../backend/graphqlschema'
 import { parseSearchURLQuery, SearchOptions } from '../search/index'
-import { SearchFilterChips } from '../search/input/SearchFilterChips'
 import { SearchNavbarItem } from '../search/input/SearchNavbarItem'
-import { eventLogger } from '../tracking/eventLogger'
 import { NavLinks } from './NavLinks'
 
 interface Props {
@@ -23,14 +21,11 @@ interface Props {
 }
 
 interface State {
-    showScopes: boolean
     authRequired?: boolean
 }
 
-const SHOW_SCOPES_LOCAL_STORAGE_KEY = 'show-scopes'
-
 export class GlobalNavbar extends React.PureComponent<Props, State> {
-    public state: State = { showScopes: false }
+    public state: State = {}
 
     private subscriptions = new Subscription()
 
@@ -42,26 +37,16 @@ export class GlobalNavbar extends React.PureComponent<Props, State> {
          */
         const options = parseSearchURLQuery(props.location.search || '')
         if (options) {
-            this.state = {
-                showScopes: localStorage.getItem(SHOW_SCOPES_LOCAL_STORAGE_KEY) === 'true',
-            }
             props.onNavbarQueryChange(options.query)
         } else {
             // If we have no component state, then we may have gotten unmounted during a route change.
             const state: SearchOptions | undefined = props.location.state
-            this.state = {
-                showScopes: localStorage.getItem(SHOW_SCOPES_LOCAL_STORAGE_KEY) === 'true',
-            }
             props.onNavbarQueryChange(state ? state.query : '')
         }
     }
 
     public componentDidMount(): void {
         this.subscriptions.add(authRequired.subscribe(authRequired => this.setState({ authRequired })))
-    }
-
-    public componentDidUpdate(): void {
-        localStorage.setItem(SHOW_SCOPES_LOCAL_STORAGE_KEY, this.state.showScopes + '')
     }
 
     public componentWillUnmount(): void {
@@ -92,35 +77,10 @@ export class GlobalNavbar extends React.PureComponent<Props, State> {
                         </div>
                     )}
                     {!this.state.authRequired && (
-                        <NavLinks
-                            {...this.props}
-                            className="global-navbar__nav-links"
-                            onShowScopes={this.onShowScopes}
-                            showScopes={this.state.showScopes}
-                            adjacentToQueryInput={true}
-                        />
+                        <NavLinks {...this.props} className="global-navbar__nav-links" adjacentToQueryInput={true} />
                     )}
                 </div>
-                {!this.state.authRequired && (
-                    <div
-                        className={
-                            'global-navbar__scopesbar' +
-                            (this.state.showScopes ? '' : ' global-navbar__scopesbar--hidden')
-                        }
-                    >
-                        <SearchFilterChips
-                            location={this.props.location}
-                            history={this.props.history}
-                            query={this.props.navbarSearchQuery}
-                        />
-                    </div>
-                )}
             </div>
         )
-    }
-
-    private onShowScopes = () => {
-        eventLogger.log('ScopesBarToggled', { code_search: { scopes_bar_toggled: !this.state.showScopes } })
-        this.setState(({ showScopes }) => ({ showScopes: !showScopes }))
     }
 }